Time: 4 ms (5.01%), Space: 25.7 MB (5.21%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -29,11 +29,11 @@ public:
         }
         vector<double> ans;
         for (auto level : result) {
-            long long sum = 0;
+            double sum = 0;
             for (auto x : level) {
                 sum += x;
             }
-            double avg = (double)sum / level.size();
+            double avg = sum / level.size();
             ans.push_back(avg);
         }
         return ans;
